requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
typing_extensions>=4.5.0

